No use inner class as interceptor type
