+++ agent/agent-plugins/mongodb/src/main/java/org/bithon/agent/plugin/mongodb/interceptor/CommandHelper$ExecuteCommandAsync.java
@@ -1,0 +1,35 @@
+/*
+ *    Copyright 2020 bithon.org
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package org.bithon.agent.plugin.mongodb.interceptor;
+
+import com.mongodb.async.SingleResultCallback;
+import org.bithon.agent.instrumentation.aop.context.AopContext;
+import org.bithon.agent.instrumentation.aop.interceptor.declaration.AfterInterceptor;
+import org.bson.BsonDocument;
+
+/**
+ * {@link com.mongodb.connection.CommandHelper#executeCommandAsync(String, BsonDocument, com.mongodb.connection.InternalConnection, SingleResultCallback)}
+ *
+ * @author frank.chen021@outlook.com
+ * @date 2021/3/29 1:53 下午
+ */
+public class CommandHelper$ExecuteCommandAsync extends AfterInterceptor {
+    @Override
+    public void after(AopContext aopContext) throws Exception {
+        super.after(aopContext);
+    }
+}
