+++ agent/agent-plugins/mongodb/src/main/java/org/bithon/agent/plugin/mongodb/interceptor/DeleteProtocol$Ctor.java
@@ -1,0 +1,44 @@
+/*
+ *    Copyright 2020 bithon.org
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package org.bithon.agent.plugin.mongodb.interceptor;
+
+
+import com.mongodb.MongoNamespace;
+import com.mongodb.WriteConcern;
+import org.bithon.agent.instrumentation.aop.IBithonObject;
+import org.bithon.agent.instrumentation.aop.context.AopContext;
+import org.bithon.agent.instrumentation.aop.interceptor.declaration.AfterInterceptor;
+import org.bithon.agent.observability.metric.domain.mongo.MongoCommand;
+
+import java.util.List;
+
+/**
+ * {@link com.mongodb.connection.DeleteProtocol#DeleteProtocol(MongoNamespace, boolean, WriteConcern, List)}
+ *
+ * @author frank.chen021@outlook.com
+ * @date 2021/3/28 11:40
+ */
+public class DeleteProtocol$Ctor extends AfterInterceptor {
+    @Override
+    public void after(AopContext aopContext) {
+        MongoNamespace ns = aopContext.getArgAs(0);
+        IBithonObject bithonObject = aopContext.getTargetAs();
+        bithonObject.setInjectedObject(new MongoCommand(ns.getDatabaseName(),
+                                                        ns.getCollectionName(),
+                                                        "Delete"));
+    }
+}
